feat: Slots are working
-> We can add slots now
-> endppint for slots is secure is jwt

+++ src/main/java/com/example/restaurantbooking/repository/SlotRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SlotRepository extends JpaRepository<Slot, Integer> {
+public interface SlotRepository extends JpaRepository<Slot, Long> {
     List<Slot> findByRestaurant(Restaurant restaurant);
 }
